refactoring : #40 - 테스트 케이스 개선
